requirements.txt
# Test
pytest>=7.0.0

# Ingestion (LangChain)
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# Vector storage (LangChain Chroma integration)
chromadb>=0.4.0
langchain-chroma>=0.1.0

# LLM (optional; use mock if OPENAI_API_KEY not set)
langchain-openai>=0.2.0

# Environment (load .env for OPENAI_API_KEY, etc.)
python-dotenv>=1.0.0

# Web: Streamlit + API
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
